Data.Map.Strict.Expiring

Documentation

data Map g k a

A map of values that expire after a given generation.

new :: g -> Map g k a

Make a new empty Map at the starting generation.

generation :: Map g k a -> g

The current generation

newGen :: (Ord k, Ord g) => g -> Map g k a -> Map g k a

𝑂(log𝑛). Assign the next generation and expire any data this new generation invalidates. The generation may never decrease. Attempts to decrease it are ignored.

insert :: (Ord k, Ord g) => g -> k -> a -> Map g k a -> Map g k a

𝑂(log𝑛). Insert a new value into the map to expire after the given generation. alterF :: (Functor f, Ord k) => (Maybe a -> f (Maybe a)) -> k -> Map k a -> f (Map k a)

delete :: (Ord k, Ord g) => k -> Map g k a -> Map g k a

𝑂(log𝑛). Delete an item.

lookup :: (Ord k, Ord g) => k -> Map g k a -> Maybe a

𝑂(log𝑛). Lookup a value in the map. This will not return any items that have expired.

updateLookupWithKey :: (Ord g, Ord k) => g -> (k -> a -> Maybe a) -> k -> Map g k a -> (Maybe a, Map g k a)

𝑂(log𝑛). Lookup and update. The function returns changed value, if it is updated. Returns the original key value if the map entry is deleted.

assocs :: Ord g => Map g k a -> [(k, a)]

𝑂(𝑛). Return all current key/value associations.

for testing

inspect :: Ord k => Map g k a -> (Int, g, Int)

Inspect stored size for testing.