Network.MQTT.Arbitrary

Description

Arbitrary instances for QuickCheck.

Documentation

newtype SizeT

Arbitrary type fitting variable integers.

Constructors

SizeT Int

newtype MatchingTopic

An arbitrary Topic and an arbitrary Filter that should match it.

Constructors

MatchingTopic (Topic, [Filter])

arbitraryTopicSegment :: [Char] -> Int -> Gen Text

Generate an arbitrary topic segment (e.g. the X in 'a/X/b') of a given length from the given alphabet.

arbitraryTopic :: [Char] -> (Int, Int) -> (Int, Int) -> Gen Topic

Generate an arbitrary Topic from the given alphabet with lengths of segments and the segment count specified by the given ranges.

arbitraryFilter :: [Char] -> (Int, Int) -> (Int, Int) -> Gen Filter

Generate an arbitrary Filter from the given alphabet with lengths of segments and the segment count specified by the given ranges. Segments may contain wildcards.

arbitraryMatchingTopic :: [Char] -> (Int, Int) -> (Int, Int) -> (Int, Int) -> Gen (Topic, [Filter])

Generate an arbitrary topic similarly to arbitraryTopic as well as some arbitrary filters that should match that topic.

v311mask :: MQTTPkt -> MQTTPkt

v311mask strips all the v5 specific bits from an MQTTPkt.

Orphan instances

Arbitrary Filter

Instance details

Methods

arbitrary :: Gen Filter

shrink :: Filter -> [Filter]

Arbitrary Topic

Instance details

Methods

arbitrary :: Gen Topic

shrink :: Topic -> [Topic]

Arbitrary AuthRequest

Instance details

Methods

arbitrary :: Gen AuthRequest

shrink :: AuthRequest -> [AuthRequest]

Arbitrary ConnACKFlags

Instance details

Methods

arbitrary :: Gen ConnACKFlags

shrink :: ConnACKFlags -> [ConnACKFlags]

Arbitrary ConnACKRC

Instance details

Methods

arbitrary :: Gen ConnACKRC

shrink :: ConnACKRC -> [ConnACKRC]

Arbitrary ConnectRequest

Instance details

Methods

arbitrary :: Gen ConnectRequest

shrink :: ConnectRequest -> [ConnectRequest]

Arbitrary DiscoReason

Instance details

Methods

arbitrary :: Gen DiscoReason

shrink :: DiscoReason -> [DiscoReason]

Arbitrary DisconnectRequest

Instance details

Methods

arbitrary :: Gen DisconnectRequest

shrink :: DisconnectRequest -> [DisconnectRequest]

Arbitrary LastWill

Instance details

Methods

arbitrary :: Gen LastWill

shrink :: LastWill -> [LastWill]

Arbitrary MQTTPkt

Instance details

Methods

arbitrary :: Gen MQTTPkt

shrink :: MQTTPkt -> [MQTTPkt]

Arbitrary Property

Instance details

Methods

arbitrary :: Gen Property

shrink :: Property -> [Property]

Arbitrary ProtocolLevel

Instance details

Methods

arbitrary :: Gen ProtocolLevel

shrink :: ProtocolLevel -> [ProtocolLevel]

Arbitrary PubACK

Instance details

Methods

arbitrary :: Gen PubACK

shrink :: PubACK -> [PubACK]

Arbitrary PubCOMP

Instance details

Methods

arbitrary :: Gen PubCOMP

shrink :: PubCOMP -> [PubCOMP]

Arbitrary PubREC

Instance details

Methods

arbitrary :: Gen PubREC

shrink :: PubREC -> [PubREC]

Arbitrary PubREL

Instance details

Methods

arbitrary :: Gen PubREL

shrink :: PubREL -> [PubREL]

Arbitrary PublishRequest

Instance details

Methods

arbitrary :: Gen PublishRequest

shrink :: PublishRequest -> [PublishRequest]

Arbitrary QoS

Instance details

Methods

arbitrary :: Gen QoS

shrink :: QoS -> [QoS]

Arbitrary SessionReuse

Instance details

Methods

arbitrary :: Gen SessionReuse

shrink :: SessionReuse -> [SessionReuse]

Arbitrary SubErr

Instance details

Methods

arbitrary :: Gen SubErr

shrink :: SubErr -> [SubErr]

Arbitrary SubOptions

Instance details

Methods

arbitrary :: Gen SubOptions

shrink :: SubOptions -> [SubOptions]

Arbitrary SubscribeRequest

Instance details

Methods

arbitrary :: Gen SubscribeRequest

shrink :: SubscribeRequest -> [SubscribeRequest]

Arbitrary SubscribeResponse

Instance details

Methods

arbitrary :: Gen SubscribeResponse

shrink :: SubscribeResponse -> [SubscribeResponse]

Arbitrary UnsubStatus

Instance details

Methods

arbitrary :: Gen UnsubStatus

shrink :: UnsubStatus -> [UnsubStatus]

Arbitrary UnsubscribeRequest

Instance details

Methods

arbitrary :: Gen UnsubscribeRequest

shrink :: UnsubscribeRequest -> [UnsubscribeRequest]

Arbitrary UnsubscribeResponse

Instance details

Methods

arbitrary :: Gen UnsubscribeResponse

shrink :: UnsubscribeResponse -> [UnsubscribeResponse]