Network.MQTT.Client

Description

An MQTT protocol client

Both MQTT 3.1.1 and MQTT 5.0 are supported over plain TCP, TLS, WebSockets and Secure WebSockets.

Configuring the client.

data MQTTConfig

Configuration for setting up an MQTT client.

Constructors

MQTTConfig

Fields

_cleanSession :: Bool

False if a session should be reused.

_lwt :: Maybe LastWill

LastWill message to be sent on client disconnect.

_msgCB :: MessageCallback

Callback for incoming messages.

_protocol :: ProtocolLevel

Protocol to use for the connection.

_connProps :: [Property]

Properties to send to the broker in the CONNECT packet.

_hostname :: String

Host to connect to (parsed from the URI)

_port :: Int

Port number (parsed from the URI)

_connID :: String

Unique connection ID (parsed from the URI)

_username :: Maybe String

Optional username (parsed from the URI)

_password :: Maybe String

Optional password (parsed from the URI)

_connectTimeout :: Int

Connection timeout (microseconds)

_tlsSettings :: TLSSettings

TLS Settings for secure connections

data MQTTClient

The MQTT client.

See connectURI for the most straightforward example.

data QoS

QoS values for publishing and subscribing.

Constructors

QoS0

QoS1

QoS2

data Topic

An MQTT topic.

mqttConfig :: MQTTConfig

A default MQTTConfig. A _connID may be required depending on your broker (or if you just want an identifiable/resumable connection). In MQTTv5, an empty connection ID may be sent and the server may assign an identifier for you and return it in the PropAssignedClientIdentifier Property.

mkLWT :: Topic -> ByteString -> Bool -> LastWill

A convenience method for creating a LastWill.

data LastWill

An MQTT Will message.

Constructors

LastWill

Fields

_willRetain :: Bool

_willQoS :: QoS

_willTopic :: ByteString

_willMsg :: ByteString

_willProps :: [Property]

data ProtocolLevel

MQTT Protocol Levels

Constructors

Protocol311
MQTT 3.1.1

Protocol50
MQTT 5.0

data Property

Property represents the various MQTT Properties that may sent or received along with packets in MQTT 5. For detailed use on when and where to use them, consult with the MQTT 5.0 spec.

Constructors

PropPayloadFormatIndicator Word8

PropMessageExpiryInterval Word32

PropContentType ByteString

PropResponseTopic ByteString

PropCorrelationData ByteString

PropSubscriptionIdentifier Int

PropSessionExpiryInterval Word32

PropAssignedClientIdentifier ByteString

PropServerKeepAlive Word16

PropAuthenticationMethod ByteString

PropAuthenticationData ByteString

PropRequestProblemInformation Word8

PropWillDelayInterval Word32

PropRequestResponseInformation Word8

PropResponseInformation ByteString

PropServerReference ByteString

PropReasonString ByteString

PropReceiveMaximum Word16

PropTopicAliasMaximum Word16

PropTopicAlias Word16

PropMaximumQoS Word8

PropRetainAvailable Word8

PropUserProperty ByteString ByteString

PropMaximumPacketSize Word32

PropWildcardSubscriptionAvailable Word8

PropSubscriptionIdentifierAvailable Word8

PropSharedSubscriptionAvailable Word8

data SubOptions

Options used at subscribe time to define how to handle incoming messages.

Constructors

SubOptions

Fields

_retainHandling :: RetainHandling

How to handle existing retained messages.

_retainAsPublished :: Bool

If true, retain is propagated on subscribe.

_noLocal :: Bool

If true, do not send messages initiated from this client back.

_subQoS :: QoS

Maximum QoS to use for this subscription.

subOptions :: SubOptions

Reasonable subscription option defaults at QoS0.

data MessageCallback

Callback invoked on each incoming subscribed message.

Constructors

NoCallback

SimpleCallback (MQTTClient -> Topic -> ByteString -> [Property] -> IO ())
Callbacks will be invoked asynchronously, ordering is likely preserved, but not guaranteed. In high throughput scenarios, slow callbacks may result in a high number of Haskell threads, potentially bringing down the entire application when running out of memory. Typically faster than OrderedCallback.

OrderedCallback (MQTTClient -> Topic -> ByteString -> [Property] -> IO ())
Callbacks are guaranteed to be invoked in the same order messages are received. In high throughput scenarios, slow callbacks may cause the underlying TCP connection to block, potentially being terminated by the broker. Typically slower than SimpleCallback.

LowLevelCallback (MQTTClient -> PublishRequest -> IO ())
A LowLevelCallback receives the client and the entire publish request, providing access to all of the fields of the request. This is slightly harder to use than SimpleCallback for common cases, but there are cases where you need all the things.

OrderedLowLevelCallback (MQTTClient -> PublishRequest -> IO ())
A low level callback that is ordered.

Running and waiting for the client.

waitForClient :: MQTTClient -> IO ()

Wait for a client to terminate its connection. An exception is thrown if the client didn't terminate expectedly.

connectURI :: MQTTConfig -> URI -> IO MQTTClient

Connect to an MQTT server by URI.

mqtt://, mqtts://, ws://, and wss:// URLs are supported. The host, port, username, and password will be derived from the URI and the values supplied in the config will be ignored.

main :: IO
main = do
  let (Just uri) = parseURI "mqtt://test.mosquitto.org"
  mc <- connectURI mqttConfig{} uri
  publish mc "tmp/topic" "hello!" False

isConnected :: MQTTClient -> IO Bool

True if we're currently in a normally connected state (in the IO monad).

disconnect :: MQTTClient -> DiscoReason -> [Property] -> IO ()

Disconnect from the MQTT server.

normalDisconnect :: MQTTClient -> IO ()

Disconnect with DiscoNormalDisconnection and no properties.

General client interactions.

subscribe :: MQTTClient -> [(Filter, SubOptions)] -> [Property] -> IO ([Either SubErr QoS], [Property])

Subscribe to a list of topic filters with their respective QoSes. The accepted QoSes are returned in the same order as requested.

unsubscribe :: MQTTClient -> [Filter] -> [Property] -> IO ([UnsubStatus], [Property])

Unsubscribe from a list of topic filters.

In MQTT 3.1.1, there is no body to an unsubscribe response, so it can be ignored. If this returns, you were unsubscribed. In MQTT 5, you'll get a list of unsub status values corresponding to your request filters, and whatever properties the server thought you should know about.

publish

Arguments

:: MQTTClient

-> Topic
Topic

-> ByteString
Message body

-> Bool
Retain flag

-> IO ()

Publish a message (QoS 0).

publishq

Arguments

:: MQTTClient

-> Topic
Topic

-> ByteString
Message body

-> Bool
Retain flag

-> QoS
QoS

-> [Property]
Properties

-> IO ()

Publish a message with the specified QoS and Properties list.

pubAliased

Arguments

:: MQTTClient

-> Topic
Topic

-> ByteString
Message body

-> Bool
Retain flag

-> QoS
QoS

-> [Property]
Properties

-> IO ()

Publish a message with the specified QoS and Property list. If possible, use an alias to shorten the message length. The alias list is managed by the client in a first-come, first-served basis, so if you use this with more properties than the broker allows, only the first N (up to TopicAliasMaximum, as specified by the broker at connect time) will be aliased.

This is safe to use as a general publish mechanism, as it will default to not aliasing whenver there's not already an alias and we can't create any more.

svrProps :: MQTTClient -> IO [Property]

Get the list of properties that were sent from the broker at connect time.

connACK :: MQTTClient -> IO ConnACKFlags

Get the complete connection aCK packet from the beginning of this session.

data MQTTException

Constructors

Timeout

BadData

Discod DisconnectRequest

MQTTException String

Low-level bits

runMQTTConduit

Arguments

:: ((MQTTConduit -> IO ()) -> IO ())
an action providing an MQTTConduit in an execution context

-> MQTTConfig
the MQTTConfig

-> IO MQTTClient

Set up and run a client with a conduit context function.

The provided action calls another IO action with a MQTTConduit as a parameter. It is expected that this action will manage the lifecycle of the conduit source/sink on behalf of the client.

type MQTTConduit = (ConduitT () ByteString IO (), ConduitT ByteString Void IO ())

MQTTConduit provides a source and sink for data as used by runMQTTConduit.

isConnectedSTM :: MQTTClient -> STM Bool

True if we're currently in a normally connected state (in the STM monad).

connACKSTM :: MQTTClient -> STM ConnACKFlags

Get the complete connection ACK packet from the beginning of this session.

registerCorrelated :: MQTTClient -> ByteString -> MessageCallback -> STM ()

Register a callback handler for a message with the given correlated data identifier.

This registration will remain in place until unregisterCorrelated is called to remove it.

unregisterCorrelated :: MQTTClient -> ByteString -> STM ()

Unregister a callback handler for the given correlated data identifier.