Network.MQTT.Topic

Description

Topic and topic related utiilities.

Documentation

data Filter

An MQTT topic filter.

unFilter :: Filter -> Text

data Topic

An MQTT topic.

unTopic :: Topic -> Text

match :: Filter -> Topic -> Bool

match returns true iff the given pattern can be matched by the specified Topic as defined in the MQTT 3.1.1 specification.

mkFilter :: Text -> Maybe Filter

mkFilter creates a filter from a text representation of a valid filter.

mkTopic :: Text -> Maybe Topic

mkTopic creates a topic from a text representation of a valid filter.

split :: Splittable a => a -> [a]

split separates a Filter or Topic into its /-separated components.

toFilter :: Topic -> Filter

Convert a Topic to a Filter as all Topics are valid Filters