Network.MQTT.Types

Description

MQTT Types.

Documentation

data LastWill

An MQTT Will message.

Constructors

LastWill

Fields

_willRetain :: Bool

_willQoS :: QoS

_willTopic :: ByteString

_willMsg :: ByteString

_willProps :: [Property]

data MQTTPkt

Constructors

ConnPkt ConnectRequest ProtocolLevel

ConnACKPkt ConnACKFlags

PublishPkt PublishRequest

PubACKPkt PubACK

PubRECPkt PubREC

PubRELPkt PubREL

PubCOMPPkt PubCOMP

SubscribePkt SubscribeRequest

SubACKPkt SubscribeResponse

UnsubscribePkt UnsubscribeRequest

UnsubACKPkt UnsubscribeResponse

PingPkt

PongPkt

DisconnectPkt DisconnectRequest

AuthPkt AuthRequest

data QoS

QoS values for publishing and subscribing.

Constructors

QoS0

QoS1

QoS2

qosFromInt :: Int -> Maybe QoS

Parse a valid QoS value from an Int.

data ConnectRequest

Constructors

ConnectRequest

Fields

_username :: Maybe ByteString

_password :: Maybe ByteString

_lastWill :: Maybe LastWill

_cleanSession :: Bool

_keepAlive :: Word16

_connID :: ByteString

_connProperties :: [Property]

connectRequest :: ConnectRequest

data SessionReuse

Constructors

NewSession

ExistingSession

data ConnACKFlags

Connection acknowledgment details.

Constructors

ConnACKFlags SessionReuse ConnACKRC [Property]

data ConnACKRC

Constructors

ConnAccepted

UnacceptableProtocol

IdentifierRejected

ServerUnavailable

BadCredentials

NotAuthorized

ConnUnspecifiedError

ConnMalformedPacket

ConnProtocolError

ConnImplementationSpecificError

ConnUnsupportedProtocolVersion

ConnClientIdentifierNotValid

ConnBadUserNameOrPassword

ConnNotAuthorized

ConnServerUnavailable

ConnServerBusy

ConnBanned

ConnBadAuthenticationMethod

ConnTopicNameInvalid

ConnPacketTooLarge

ConnQuotaExceeded

ConnPayloadFormatInvalid

ConnRetainNotSupported

ConnQosNotSupported

ConnUseAnotherServer

ConnServerMoved

ConnConnectionRateExceeded

data PublishRequest

Constructors

PublishRequest

Fields

_pubDup :: Bool

_pubQoS :: QoS

_pubRetain :: Bool

_pubTopic :: ByteString

_pubPktID :: PktID

_pubBody :: ByteString

_pubProps :: [Property]

data PubACK

Constructors

PubACK PktID Word8 [Property]

data PubREC

Constructors

PubREC PktID Word8 [Property]

data PubREL

Constructors

PubREL PktID Word8 [Property]

data PubCOMP

Constructors

PubCOMP PktID Word8 [Property]

data ProtocolLevel

MQTT Protocol Levels

Constructors

Protocol311
MQTT 3.1.1

Protocol50
MQTT 5.0

data Property

Property represents the various MQTT Properties that may sent or received along with packets in MQTT 5. For detailed use on when and where to use them, consult with the MQTT 5.0 spec.

Constructors

PropPayloadFormatIndicator Word8

PropMessageExpiryInterval Word32

PropContentType ByteString

PropResponseTopic ByteString

PropCorrelationData ByteString

PropSubscriptionIdentifier Int

PropSessionExpiryInterval Word32

PropAssignedClientIdentifier ByteString

PropServerKeepAlive Word16

PropAuthenticationMethod ByteString

PropAuthenticationData ByteString

PropRequestProblemInformation Word8

PropWillDelayInterval Word32

PropRequestResponseInformation Word8

PropResponseInformation ByteString

PropServerReference ByteString

PropReasonString ByteString

PropReceiveMaximum Word16

PropTopicAliasMaximum Word16

PropTopicAlias Word16

PropMaximumQoS Word8

PropRetainAvailable Word8

PropUserProperty ByteString ByteString

PropMaximumPacketSize Word32

PropWildcardSubscriptionAvailable Word8

PropSubscriptionIdentifierAvailable Word8

PropSharedSubscriptionAvailable Word8

data AuthRequest

Constructors

AuthRequest Word8 [Property]

data SubscribeRequest

Constructors

SubscribeRequest PktID [(ByteString, SubOptions)] [Property]

data SubOptions

Options used at subscribe time to define how to handle incoming messages.

Constructors

SubOptions

Fields

_retainHandling :: RetainHandling

How to handle existing retained messages.

_retainAsPublished :: Bool

If true, retain is propagated on subscribe.

_noLocal :: Bool

If true, do not send messages initiated from this client back.

_subQoS :: QoS

Maximum QoS to use for this subscription.

subOptions :: SubOptions

Reasonable subscription option defaults at QoS0.

data SubscribeResponse

Constructors

SubscribeResponse PktID [Either SubErr QoS] [Property]

data SubErr

Constructors

SubErrUnspecifiedError

SubErrImplementationSpecificError

SubErrNotAuthorized

SubErrTopicFilterInvalid

SubErrPacketIdentifierInUse

SubErrQuotaExceeded

SubErrSharedSubscriptionsNotSupported

SubErrSubscriptionIdentifiersNotSupported

SubErrWildcardSubscriptionsNotSupported

data RetainHandling

How to process retained messages on subscriptions.

Constructors

SendOnSubscribe
Send existing retained messages to a new client.

SendOnSubscribeNew
Send existing retained messages that have not yet been sent.

DoNotSendOnSubscribe
Don't send existing retained messages.

data DisconnectRequest

Constructors

DisconnectRequest DiscoReason [Property]

data UnsubscribeRequest

Constructors

UnsubscribeRequest PktID [ByteString] [Property]

data UnsubscribeResponse

Constructors

UnsubscribeResponse PktID [Property] [UnsubStatus]

data UnsubStatus

Constructors

UnsubSuccess

UnsubNoSubscriptionExisted

UnsubUnspecifiedError

UnsubImplementationSpecificError

UnsubNotAuthorized

UnsubTopicFilterInvalid

UnsubPacketIdentifierInUse

data DiscoReason

Constructors

DiscoNormalDisconnection

DiscoDisconnectWithWill

DiscoUnspecifiedError

DiscoMalformedPacket

DiscoProtocolError

DiscoImplementationSpecificError

DiscoNotAuthorized

DiscoServerBusy

DiscoServershuttingDown

DiscoKeepAliveTimeout

DiscoSessiontakenOver

DiscoTopicFilterInvalid

DiscoTopicNameInvalid

DiscoReceiveMaximumExceeded

DiscoTopicAliasInvalid

DiscoPacketTooLarge

DiscoMessageRateTooHigh

DiscoQuotaExceeded

DiscoAdministrativeAction

DiscoPayloadFormatInvalid

DiscoRetainNotSupported

DiscoQoSNotSupported

DiscoUseAnotherServer

DiscoServerMoved

DiscoSharedSubscriptionsNotSupported

DiscoConnectionRateExceeded

DiscoMaximumConnectTime

DiscoSubscriptionIdentifiersNotSupported

DiscoWildcardSubscriptionsNotSupported

type PktID = Word16

parsePacket :: ProtocolLevel -> Parser MQTTPkt

class ByteMe a where

Minimal complete definition

Nothing

Methods

toByteString :: ProtocolLevel -> a -> ByteString

parseConnect :: Parser MQTTPkt

Parse a CONNect packet. This is useful when examining the beginning of the stream as it allows you to determine the protocol being used throughout the rest of the session.

encodeLength :: Int -> [Word8]

parseHdrLen :: Parser Int

parseProperty :: Parser Property

parseProperties :: ProtocolLevel -> Parser [Property]

bsProps :: ProtocolLevel -> [Property] -> ByteString

parseSubOptions :: Parser SubOptions

class ByteSize a where

Methods

toByte :: a -> Word8

fromByte :: Word8 -> a