Network.MQTT.Topic

Description

Topic and topic related utiilities.

Documentation

type Filter = Text

An MQTT topic filter.

type Topic = Text

An MQTT topic.

match :: Filter -> Topic -> Bool

match returns true iff the given pattern can be matched by the specified Topic as defined in the MQTT 3.1.1 specification.